If false
The check should be for if `parsedBody.error.message` is falsey, but
with `!!` it simply casts it to a bool, so for a valid error, it was
evaluating as truthy e.g. `false || false || true` => `true`, so
`Unknown error` was being returned when it should be branching into the
`else`. In this specific case, the raw error message was

```
'{"error":{"message":"A dataset with different fields already exists
with the ID \"base.deals09\""}}'
```

So instead of showing that error message, the very unhelpful `Unknown
error` was getting logged.